+++ CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Add options for command output handling:
+  - `--no-header`
+  - `--result` with `always`, `never` and `non-zero`
+
 ### Changed
 
 - `--list` is no longer required (and will be removing in the next major release). The default without a command is to print the found paths which is simpler to use.